equine profile page file optimisation changes
