+++ next.config.js
@@ -9,3 +9,9 @@ module.exports = withPWA({
   },
   reactStrictMode: true,
 })
+
+const withBundleAnalyzer = require("@next/bundle-analyzer")({
+	enabled: process.env.ANALYZE === "true",
+});
+
+module.exports = withBundleAnalyzer({});
