+++ package.json
@@ -22,6 +22,7 @@
 		"styled-components": "^5.3.5"
 	},
 	"devDependencies": {
+		"@next/bundle-analyzer": "^12.3.1",
 		"@testing-library/react": "^13.4.0",
 		"@types/jest": "^29.1.1",
 		"@types/node": "17.0.4",
